+++ src/pages/[id]/[details]/index.tsx
@@ -1,20 +1,20 @@
-import { NextPage } from 'next';
-import styled from '@emotion/styled';
-import Head from 'next/head';
+import { NextPage } from "next";
+import styled from "@emotion/styled";
+import Head from "next/head";
 
-import { DetailHeader } from '../../../components/Details/DetailHeader';
-import { LeftHeader } from '../../../components/Details/LeftHeader';
-import { RightHeader } from '../../../components/Details/RightHeader';
+import { DetailHeader } from "../../../components/Details/DetailHeader";
+import { LeftHeader } from "../../../components/Details/LeftHeader";
+import { RightHeader } from "../../../components/Details/RightHeader";
 
 // import { CardContainer } from "../../../components/Home/CardContainer";
-import { Header } from '../../../components/Common/Header';
+import { Header } from "../../../components/Common/Header";
 
-import { Theme } from '../../../styles/theme';
-import { useContext } from 'react';
-import { ThemeContext } from '../../_app';
-import { useRouter } from 'next/router';
-import { useData } from '../../../hooks/useData';
-import { ErrorPage } from '../../../components/Details/ErrorPage';
+import { Theme } from "../../../styles/theme";
+import { useContext } from "react";
+import { ThemeContext } from "../../_app";
+import { useRouter } from "next/router";
+import { useData } from "../../../hooks/useData";
+import { ErrorPage } from "../../../components/Details/ErrorPage";
 
 interface ThemeProps {
   theme: Theme;
@@ -80,8 +80,8 @@ const DetailsIndexPage: NextPage = () => {
   const userDetails = router.query.details;
 
   const { data: DetailData, error: DetailError } = useData(
-    'posts',
-    'populate=*'
+    "posts",
+    "populate=*"
   );
 
   if (!DetailData) return <div>로딩중</div>;
@@ -90,16 +90,16 @@ const DetailsIndexPage: NextPage = () => {
   let postid = 0;
 
   let postObj = {
-    title: '',
-    contents: '',
-    url: '',
+    title: "",
+    contents: "",
+    url: "",
     likeposts: {
       data: [],
     },
     comments: {
       data: [],
     },
-    createdAt: '',
+    createdAt: "",
   };
 
   DetailData.data.some((details: Post) => {
@@ -122,7 +122,7 @@ const DetailsIndexPage: NextPage = () => {
       </Head>
       {postObj.title ? (
         <div>
-          <Header username={'deli-ght'} user={true} />
+          <Header username={"deli-ght"} user={true} />
           <DetailContainer>
             <LeftHeader likepost={postObj.likeposts.data} />
             <DetailHeader
